# [Perf] Reduce SpaceCardsViewModel Creation