@EdwardJTL 2022-07-12
had to fix a bug with pinned spaces. 
when a space becomes pinned pinned, `indexInStore  != indexInDetails ` check in `listenManagerState` does not pass, and the array struct is not updated, so no filtering is triggered. @jonathangarelick i'm not sure if this is a bug

As a result, I added a manual trigger to fire another around of filtering to pick up the effects of `pinned` ordering.