@yozuysal 2022-07-12
thanks for digging into this Edward. Should we consider

1) reverting the sorting change as a whole
2) addressing the single instance as you have here

then looking into how we can revert the revert for the sorting PR?

Since we are not 100% happy with the current approach as well, that feature is hidden enough and wouldn't worth risking designing as we address perf issues.