@yozuysal 2022-07-12
actually I take that back. what if we use SpaceStore as an observable object here so that we react to SpaceStore.State changes?