@EdwardJTL 2022-07-12
> actually I take that back. what if we use SpaceStore as an observable object here so that we react to SpaceStore.State changes?

we could have `SpaceCardsViewModel` observe that and use `SpaceCardModel.allDetails` as a data source. but there might still be synchronization issues between when the `state` is set and when the new data is populated in `SpaceCardModel.allDetails`.

if we want to address the perf issue asap, we can just revert the sorting change and then take our time to re-design the architecture for the data flow.