@yozuysal 2022-07-12
lets prioritize addressing the perf change yeah. I am OK with reverting for now and we can revert revert later